requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bycrypt]
python-jose[cryptography]